Updating return/variable types
After running Clang’s analyzer on it there were quite a few places where there were mis-matched store sizes and signs.  I updated them so that the underlying code more closely-matches the OS APIs and the type change happens at the FUSE connection.

In this way, if FUSE ever updates the API to actually match the APIs of modern OSes, it’s a fix in one place.

Also, I updated read/write to return the total read/write value the OS returned rather than blindly return the request count.

+++ src/fs_acl.cpp
@@ -30,7 +30,7 @@ namespace fs
     bool
     dir_has_defaults(const std::string &fullpath)
     {
-      int rv;
+      ssize_t rv;
       std::string dirpath = fullpath;
 
       fs::path::dirname(dirpath);
